refactor(parser): split parser.rs into module directory by document type
Modularize the 2276-line parser.rs into focused submodules organized
by Sigma document type. No behavior changes; identical test results
(241 unit + 9 integration + 4 doc-tests pass).

New structure:
  parser/
    mod.rs          -- public API, document dispatch, shared helpers
    detection.rs    -- detection rule and detection block parsing
    correlation.rs  -- correlation rule parsing
    filter.rs       -- filter rule parsing
    tests.rs        -- all unit tests
